+++ pulse-binding-mainloop-glib/CHANGELOG.md
@@ -1,3 +1,11 @@
+# <unreleased>
+
+ * Addressed long standing todo item of linking `GMainContext` to an actual glib crate.
+    - Added dependency on `glib-sys`
+    - Replaced our own `GMainContext` type (re-exported from our `sys` crate) with the one from the
+      `glib-sys` crate.
+    - Deprecated use of `GMainContext` directly from this crate
+
 # 2.7.1 (August 13th, 2019)
 
  * Fix missed use of UB empty enum trick